code updated to include create producer template where it is called with the messageProtocol, topic config and kafka config which can now be gotten dynamically
